Abstract song progress HUD elements to share gameplay update logic
